Workaround: Prevent color bug on window add
Prevent deterministic, recursive, signal-slot hell bug in which
opening a new window that loads to a screen other than the screen
where it was invoked from causes it override the color values of
the screen from which it was invoked with the color values from
the screen where it opened.

Workaround consists of preventing old window position data from
being saved once that window is closed. It would be convenient to
remember that, but not at the cost of potentially loosing color data
as you're setting things up.